# api/openapi.yaml
openapi: 3.1.0
info:
  title: DokoSoko Control Plane API
  version: 4.0.0
  description: >-
    Control-plane contract for DokoSoko's core documentation, SDK references, credentials, recipes, reviewed tools, native tools, and MCP connector. Cookie-authenticated mutations
    also require CSRF validation. Public MCP is optional and read-only.
servers:
  - url: https://dokosoko.example.com
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/setup/begin:
    post:
      security:
        - setupToken: []
      operationId: beginSetup
      requestBody:
        $ref: '#/components/requestBodies/RootInput'
      responses:
        '201':
          $ref: '#/components/responses/MFAEnrollment'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/setup/complete:
    post:
      security: []
      operationId: completeSetup
      requestBody:
        $ref: '#/components/requestBodies/MFACompletion'
      responses:
        '201':
          $ref: '#/components/responses/AuthenticatedRoot'
  /api/v1/auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                code:
                  type: string
                  description: TOTP or unused recovery code.
              required:
                - email
                - password
                - code
      responses:
        '200':
          $ref: '#/components/responses/AuthenticatedRoot'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/v1/auth/session:
    get:
      operationId: currentSession
      responses:
        '200':
          description: Current root session.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/RootUser'
                  expires_at:
                    type: string
                    format: date-time
                required:
                  - user
                  - expires_at
        '401':
          $ref: '#/components/responses/Error'
...
  /.well-known/oauth-protected-resource/mcp:
    get:
      security: []
      operationId: getMCPProtectedResourceMetadata
      description: RFC 9728 protected-resource metadata for the singleton Deployment's Private MCP endpoint.
      responses:
        '200':
          description: Protected-resource metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtectedResourceMetadata'
        '404':
          $ref: '#/components/responses/Error'
  /oauth/register:
    post:
      security: []
      operationId: registerOAuthClient
      description: >-
        Registers a public PKCE client under RFC 7591. Identical normalized metadata returns the same durable client identifier, making network retries safe. No client secret is
        issued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthClientRegistrationRequest'
      responses:
        '201':
          description: Idempotent public-client registration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientRegistration'
        '400':
          $ref: '#/components/responses/OAuthError'
        '404':
          $ref: '#/components/responses/OAuthError'
        '429':
          $ref: '#/components/responses/OAuthRateLimit'
        '503':
          $ref: '#/components/responses/OAuthError'
  /oauth/authorize:
    get:
      security: []
      operationId: oauthAuthorize
      description: >-
        Starts authorization-code OAuth using a registered public client or HTTPS client metadata document, mandatory RFC 8707 resource binding, PKCE S256, upstream vendor OIDC,
        and fail-closed access evaluation.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            const: code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
            description: Durable RFC 7591 client identifier or HTTPS URL of a client metadata document.
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: resource
          in: query
          required: true
          schema:
            type: string
            format: uri
            description: Exact Private MCP resource URL advertised by protected-resource metadata.
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: mcp:private
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            const: S256
      responses:
        '302':
          description: Redirect to the configured vendor IdP.
        '400':
          $ref: '#/components/responses/OAuthError'
  /oauth/callback:
    get:
      security: []
      operationId: oauthCallback
      description: >-
        For public OAuth states, verifies upstream state, nonce and ID token, resolves vendor access fail-closed, creates or resolves the durable customer account, then redirects
        with a one-time code. Prefixed administration-test states stop after OIDC, audience/client, nonce, issuer, and customer-claim verification and redirect to the fixed
        same-origin Identity page without retaining tokens or creating runtime identity data.
      responses:
        '302':
          description: Exact registered client redirect.
        '303':
          description: >-
            Fixed same-origin Identity-page redirect carrying identity_test_id for a stored result or
            identity_test_error=invalid_or_expired when the callback transaction cannot be claimed.
        '401':
          $ref: '#/components/responses/OAuthError'
  /oauth/token:
    post:
      security: []
      operationId: oauthToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: false
              properties:
                grant_type:
                  const: authorization_code
                code:
                  type: string
                code_verifier:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                resource:
                  type: string
                  format: uri
              required:
                - grant_type
                - code
                - code_verifier
                - client_id
                - redirect_uri
                - resource
      responses:
        '200':
          description: Short-lived opaque DokoSoko access token bound to the exact Private MCP resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
        '400':
          $ref: '#/components/responses/OAuthError'
        '502':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /agent-setup/{kind}/prompt.md:
    parameters:
      - name: kind
        in: path
        required: true
        schema:
          type: string
          enum:
            - public
            - private
    get:
      security: []
      operationId: getAgentSetupPrompt
      description: >-
        Secret-free, client-specific setup instructions derived from the canonical configured public origin. Public setup is unavailable while Public MCP is disabled; private setup
        is unavailable while customer identity is inactive.
      responses:
        '200':
          description: Setup prompt for Codex, Claude Code, Cursor, and OpenCode.
          content:
            text/markdown:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/Error'
  /mcp/public:
    parameters:
      - $ref: '#/components/parameters/MCPProtocolVersion'
      - $ref: '#/components/parameters/MCPMethod'
      - $ref: '#/components/parameters/MCPName'
    post:
      security: []
      operationId: publicDeploymentMCP
      description: Canonical authentication-free Stateless MCPv2 endpoint for the singleton Deployment. Available only while Public MCP is enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONRPCRequest'
      responses:
        '200':
          description: JSON-RPC result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCResponse'
        '404':
          description: Public MCP is disabled.
        '429':
          description: Anonymous request budget exceeded.
  /mcp:
    parameters:
      - $ref: '#/components/parameters/MCPProtocolVersion'
      - $ref: '#/components/parameters/MCPMethod'
      - $ref: '#/components/parameters/MCPName'
    post:
      security:
        - dokoBearer: []
      operationId: privateDeploymentMCP
      description: Canonical authenticated Stateless MCPv2 endpoint for the singleton Deployment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONRPCRequest'
      responses:
        '200':
          description: JSON-RPC result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dokosoko_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator automation token in development only; production uses the root session cookie.
    setupToken:
      type: http
      scheme: bearer
      description: One-time deployment setup token.
    dokoBearer:
      type: http
      scheme: bearer
      description: Opaque, short-lived DokoSoko OAuth access token bound to one exact Private MCP resource.
  parameters:
...
    MCPProtocolVersion:
      name: MCP-Protocol-Version
      in: header
      required: true
      schema:
        type: string
        const: '2026-07-28'
    MCPMethod:
      name: Mcp-Method
      in: header
      required: true
      description: Must exactly match the JSON-RPC method.
      schema:
        type: string
        enum:
          - server/discover
          - tools/list
          - tools/call
    MCPName:
      name: Mcp-Name
      in: header
      required: false
      description: Required for tools/call and must exactly match params.name.
      schema:
        type: string
...
  requestBodies:
...
    RootInput:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              email:
                type: string
                format: email
              display_name:
                type: string
                minLength: 1
                maxLength: 120
              password:
                type: string
                format: password
                minLength: 14
            required:
              - email
              - display_name
              - password
    MFACompletion:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              enrollment_id:
                type: string
              code:
                type: string
            required:
              - enrollment_id
              - code
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    OAuthError:
      description: OAuth 2.0 error response.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              error:
                type: string
              error_description:
                type: string
            required:
              - error
              - error_description
    OAuthRateLimit:
      description: OAuth client registration rate limit exceeded. Retrying after the advertised delay is safe.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              error:
                const: temporarily_unavailable
              error_description:
                type: string
            required:
              - error
              - error_description
    MFAEnrollment:
      description: Short-lived MFA enrollment; the TOTP secret is returned only during enrollment.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MFAEnrollment'
    AuthenticatedRoot:
      description: Authenticated root session. Recovery codes are returned only once when a root is created.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              user:
                $ref: '#/components/schemas/RootUser'
              expires_at:
                type: string
                format: date-time
              csrf_token:
                type: string
                description: Double-submit CSRF token also set in the dokosoko_csrf cookie.
              recovery_codes:
                type: array
                items:
                  type: string
            required:
              - user
...
  schemas:
...
    OAuthClientRegistrationRequest:
      type: object
      additionalProperties: true
      properties:
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
        client_name:
          type: string
          maxLength: 200
        token_endpoint_auth_method:
          type: string
          const: none
          default: none
        grant_types:
          type: array
          maxItems: 1
          items:
            const: authorization_code
        response_types:
          type: array
          maxItems: 1
          items:
            const: code
        scope:
          type: string
          const: mcp:private
      required:
        - redirect_uris
    OAuthClientRegistration:
      type: object
      additionalProperties: false
      properties:
        client_id:
          type: string
        client_id_issued_at:
          type: integer
          format: int64
        client_name:
          type: string
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
        token_endpoint_auth_method:
          type: string
          const: none
        grant_types:
          type: array
          items:
            const: authorization_code
        response_types:
          type: array
          items:
            const: code
        scope:
          type: string
          const: mcp:private
      required:
        - client_id
        - client_id_issued_at
        - client_name
        - redirect_uris
        - token_endpoint_auth_method
        - grant_types
        - response_types
        - scope
    OAuthProtectedResourceMetadata:
      type: object
      additionalProperties: true
      properties:
        resource:
          type: string
          format: uri
        authorization_servers:
          type: array
          minItems: 1
          items:
            type: string
            format: uri
        scopes_supported:
          type: array
          items:
            type: string
          contains:
            const: mcp:private
        bearer_methods_supported:
          type: array
          items:
            type: string
          contains:
            const: header
      required:
        - resource
        - authorization_servers
        - scopes_supported
        - bearer_methods_supported
    OAuthTokenResponse:
      type: object
      additionalProperties: false
      properties:
        access_token:
          type: string
          description: Opaque resource-bound DokoSoko access token.
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          minimum: 1
        scope:
          type: string
      required:
        - access_token
        - token_type
        - expires_in
        - scope
    JSONRPCID:
      type:
        - string
        - integer
        - 'null'
    JSONRPCToolCallParams:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        arguments:
          type: object
          additionalProperties: true
        _meta:
          allOf:
            - $ref: '#/components/schemas/MCPRequestMeta'
            - type: object
              properties:
                confirmed:
                  type: boolean
                  default: false
                idempotency_key:
                  type: string
                  minLength: 16
                  maxLength: 200
                  pattern: ^[!-~]{16,200}$
                  description: >-
                    Stable identifier for one tool invocation. Required for custom HTTP mutations whose discovery metadata declares com.dokosoko/idempotencyKeyRequired. Reuse it
                    across transport retries; do not substitute a request ID.
      required:
        - name
        - arguments
        - _meta
    MCPRequestMeta:
      type: object
      additionalProperties: true
      properties:
        io.modelcontextprotocol/protocolVersion:
          type: string
          const: '2026-07-28'
      required:
        - io.modelcontextprotocol/protocolVersion
    JSONRPCError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: integer
        message:
          type: string
        data: {}
      required:
        - code
        - message
    JSONRPCResponse:
      type: object
      additionalProperties: false
      properties:
        jsonrpc:
          const: '2.0'
        id:
          $ref: '#/components/schemas/JSONRPCID'
        result: {}
        error:
          $ref: '#/components/schemas/JSONRPCError'
      required:
        - jsonrpc
        - id
      oneOf:
        - required:
            - result
          not:
            required:
              - error
        - required:
            - error
          not:
            required:
              - result
    RootUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          const: root
        revoked_at:
          type:
            - string
            - 'null'
          format: date-time
      required:
        - id
        - email
        - display_name
        - role
    MFAEnrollment:
      type: object
      properties:
        enrollment_id:
          type: string
        totp_secret:
          type: string
        provisioning_uri:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
      required:
        - enrollment_id
        - totp_secret
        - provisioning_uri
        - expires_at
...
    JSONRPCRequest:
      type: object
      additionalProperties: false
      properties:
        jsonrpc:
          const: '2.0'
        id:
          $ref: '#/components/schemas/JSONRPCID'
        method:
          type: string
          enum:
            - server/discover
            - tools/list
            - tools/call
            - resources/list
            - resources/read
        params:
          type: object
          additionalProperties: true
          properties:
            _meta:
              $ref: '#/components/schemas/MCPRequestMeta'
          required:
            - _meta
      required:
        - jsonrpc
        - method
        - params
      allOf:
        - if:
            properties:
              method:
                const: tools/call
          then:
            properties:
              params:
                $ref: '#/components/schemas/JSONRPCToolCallParams'
            required:
              - params
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - 'null'
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error
...